Fix DFU mode issue
Issue was due to the fact that sometimes the code remained blocked
inside the timer callback after the closure of the USB serial port.
In order to avoid that, some checks inside the timer callbeck and
the CDC_Flush API have been added.

Improving the calculation of the period and prescaler of the timer.

Signed-off-by: Carlo.Parata <carlo.parata@st.com>